Add files via upload
This is a basic game engine made in C made for quizzes, the quiz is made in a text file which is accessed through parameters and is then displayed. The player can answer the questions and the engine will tell the player if they got the answers right or wrong. there is a sample quiz in the zip file